+++ theme-suite/branches/2.0/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release Notes
 
+## [v2.0.3 (2020-11-17)](https://svn.tigreblanc.fr/presstify-plugins/theme-suite/tags/2.0.3...v2.0.3)
+
+### Changed
+
+- `src/Partial/ArticleCardPartial.php` : Test et mise en service
+
+### Fixed 
+
+- `src/ThemeSuiteAwareTrait.php` : Correctif instanciation auto
+
 ## [v2.0.2 (2020-11-16)](https://svn.tigreblanc.fr/presstify-plugins/theme-suite/tags/2.0.2...v2.0.2)
 
 ### Changed